Organize code in the src directory
