+++ src/entities/Entity.js
